Make the trait params passing more robust
Instead of using a single package-lexical to pass the params, rework
init_meta() to take a named parameter, and fall back to the package-lexical
variable.  Additionally, clear the package-lexical variable afterwards to
keep it from sticking around, if init_meta() happens to be called
independently of (and after) a previous import.

I'm pretty sure this change is just to make me feel better :)