add "builder => sub { ... }" shortcut -- closes #4
If we're passed a 'builder' option and that option is a coderef, then define a
builder of the canonical name (e.g. builder => 1) and hook
class association to install that method as a method with the appropriate
name.

Note that we don't associate the installed builder method with the attribute
or install it with accessors; this is a bit of a special case so I _think_
we're ok to just install it at class association time.  (After all, builder
methods are normally defined and implemented independently of the attribute.)
We also don't create a method metaclass for our builder method, much less one
that's traited to mark it as an installed builder, but I'm willing to change
that for a good reason if someone asks nicely :)