add TODO tests for anon builder method creation in roles
using builders in roles and being able to exclude/alias them as
necessary when consuming them is a major win.  Because of that -- and
because that's what we'd expect to be able to do with a builder when
created in the usual fashion (aka not anon sub via us) -- we want to
create the builder as a method in the role when we define the
attribute to a role.

We don't do that quite yet :)