Copy parent's type coercions over if coercion is requested
Child types do not inherit constraints from their parent, for reasons we need
not go into here.  However, often the option 'constraint' is used to make a
minor tweak to an existing constraint where we do want the parent coercions to
still work.  (e.g. isa => File, constraint => sub { $_->stat }, coerce => 1)

For this, if simple coercion is requested (that is, the coerce option is true
and eq "1") and the parent type has coercions, we copy the coercions.  Any
errors (e.g. someone request coercions but the parent "isa" type has none) we
leave to Moose proper to blow up over.