Initial creation of role attribute trait
Until now, we were just letting one trait handle attributes defined in
both roles and classes.  For roles this would simply defer everything
until the role is applied to a class, however that also meant that
anonymous builders weren't created as methods until the role was
consumed.

Here we use a role attribute trait to do the necessary bits to create a
builder method when the role is defined in a role.